refactor: remove dead code from db/queries.rs after repository migration
Remove 13 unused database query functions that have been migrated to
repository modules, completing the cleanup of legacy direct DB access.

Updated handlers to use repositories:
- auth/oauth.rs, auth/middleware.rs
- api/analytics/usage.rs, api/analytics/org.rs, api/analytics/link.rs
- api/billing/checkout.rs, api/billing/status.rs
- api/orgs/crud.rs, api/links/create.rs, api/links/admin.rs
- api/reports/create.rs, repositories/link_repository.rs
