+++ src/db/queries.rs
@@ -1,309 +1,8 @@
-use crate::models::{BillingAccount, Link, Organization, User, user::CreateUserData};
+use crate::models::{BillingAccount, User};
 use crate::repositories::link_repository::{AdminLink, serialize_optional_int_as_bool};
 use crate::utils::now_timestamp;
 use worker::d1::D1Database;
 use worker::*;
-
-/// Get the total number of users on the instance
-pub async fn get_user_count(db: &D1Database) -> Result<i64> {
-    let stmt = db.prepare("SELECT COUNT(*) as count FROM users");
-    let result = stmt.first::<serde_json::Value>(None).await?;
-    match result {
-        Some(val) => Ok(val["count"].as_f64().unwrap_or(0.0) as i64),
-        None => Ok(0),
-    }
-}
-
-/// Create or update a user from OAuth data
-pub async fn create_or_update_user(
-    db: &D1Database,
-    data: CreateUserData,
-    org_id: &str,
-) -> Result<User> {
-    let user_id = uuid::Uuid::new_v4().to_string();
-    let now = now_timestamp();
-
-    // Try to find existing user by OAuth provider and ID
-    let stmt = db.prepare(
-        "SELECT id, email, name, avatar_url, oauth_provider, oauth_id, org_id, role, created_at
-         FROM users
-         WHERE oauth_provider = ?1 AND oauth_id = ?2",
-    );
-
-    let existing = stmt
-        .bind(&[
-            data.oauth_provider.clone().into(),
-            data.oauth_id.clone().into(),
-        ])?
-        .first::<User>(None)
-        .await?;
-
-    if let Some(user) = existing {
-        // Update existing user
-        let update_stmt = db.prepare(
-            "UPDATE users
-             SET email = ?1, name = ?2, avatar_url = ?3
-             WHERE id = ?4",
-        );
-
-        update_stmt
-            .bind(&[
-                data.email.clone().into(),
-                data.name.clone().into(),
-                data.avatar_url.clone().into(),
-                user.id.clone().into(),
-            ])?
-            .run()
-            .await?;
-
-        Ok(User {
-            id: user.id,
-            email: data.email,
-            name: data.name,
-            avatar_url: data.avatar_url,
-            oauth_provider: data.oauth_provider,
-            oauth_id: data.oauth_id,
-            org_id: user.org_id,
-            role: user.role,
-            created_at: user.created_at,
-            suspended_at: user.suspended_at,
-            suspension_reason: user.suspension_reason,
-            suspended_by: user.suspended_by,
-        })
-    } else {
-        // Determine role: first user on the instance gets admin, all others get member
-        let user_count = get_user_count(db).await?;
-        let role = if user_count == 0 { "admin" } else { "member" };
-
-        // Create new user
-        let insert_stmt = db.prepare(
-            "INSERT INTO users (id, email, name, avatar_url, oauth_provider, oauth_id, org_id, role, created_at)
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9)"
-        );
-
-        insert_stmt
-            .bind(&[
-                user_id.clone().into(),
-                data.email.clone().into(),
-                data.name.clone().into(),
-                data.avatar_url.clone().into(),
-                data.oauth_provider.clone().into(),
-                data.oauth_id.clone().into(),
-                org_id.into(),
-                role.into(),
-                (now as f64).into(),
-            ])?
-            .run()
-            .await?;
-
-        Ok(User {
-            id: user_id,
-            email: data.email,
-            name: data.name,
-            avatar_url: data.avatar_url,
-            oauth_provider: data.oauth_provider,
-            oauth_id: data.oauth_id,
-            org_id: org_id.to_string(),
-            role: role.to_string(),
-            created_at: now,
-            suspended_at: None,
-            suspension_reason: None,
-            suspended_by: None,
-        })
-    }
-}
-
-/// Generate a unique slug for an organization name
-/// Always adds random 5-character suffix for consistency and collision prevention
-pub async fn generate_unique_slug(_db: &D1Database, org_name: &str) -> Result<String> {
-    // Generate base slug using existing logic
-    let base_slug = org_name
-        .to_lowercase()
-        .replace(' ', "-")
-        .chars()
-        .filter(|c| c.is_alphanumeric() || *c == '-')
-        .collect::<String>();
-
-    let uuid_str = uuid::Uuid::new_v4().to_string().replace('-', "");
-    let random_suffix = &uuid_str[..5];
-
-    let slug = format!("{}-{}", base_slug, random_suffix);
-
-    Ok(slug)
-}
-
-#[cfg(test)]
-mod slug_tests {
-    // Test slug generation logic without database dependency
-    #[tokio::test]
-    async fn test_slug_generation_logic() {
-        // Test base slug generation
-        let base_slug = "Test Organization"
-            .to_lowercase()
-            .replace(' ', "-")
-            .chars()
-            .filter(|c| c.is_alphanumeric() || *c == '-')
-            .collect::<String>();
-
-        assert_eq!(base_slug, "test-organization");
-
-        // Test with special characters
-        let special_slug = "John's Test Organization! @#$%"
-            .to_lowercase()
-            .replace(' ', "-")
-            .chars()
-            .filter(|c| c.is_alphanumeric() || *c == '-')
-            .collect::<String>();
-
-        assert_eq!(special_slug, "johns-test-organization-");
-
-        // Test UUID-based suffix generation
-        let uuid_str = uuid::Uuid::new_v4().to_string().replace('-', "");
-        let suffix = &uuid_str[..5];
-        assert_eq!(suffix.len(), 5);
-        assert!(suffix.chars().all(|c| c.is_alphanumeric()));
-
-        // Test full slug generation (always includes suffix)
-        let full_slug = format!("{}-{}", base_slug, suffix);
-        assert!(full_slug.starts_with("test-organization-"));
-        assert_eq!(full_slug.len(), base_slug.len() + 1 + 5); // base + "-" + 5 chars
-    }
-}
-
-/// Create a default organization for a new user
-pub async fn create_default_org(
-    db: &D1Database,
-    user_id: &str,
-    org_name: &str,
-) -> Result<Organization> {
-    let org_id = uuid::Uuid::new_v4().to_string();
-    let slug = generate_unique_slug(db, org_name).await?;
-
-    let now = now_timestamp();
-
-    // Read the default tier from settings
-    let settings_repo = crate::repositories::SettingsRepository::new();
-    let tier = settings_repo
-        .get_setting(db, "default_user_tier")
-        .await?
-        .unwrap_or_else(|| "free".to_string());
-
-    // Create billing account for the user
-    let billing_account = create_billing_account(db, user_id, &tier).await?;
-
-    let stmt = db.prepare(
-        "INSERT INTO organizations (id, name, slug, created_at, created_by, billing_account_id)
-         VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-    );
-
-    stmt.bind(&[
-        org_id.clone().into(),
-        org_name.to_string().into(),
-        slug.clone().into(),
-        (now as f64).into(),
-        user_id.into(),
-        billing_account.id.clone().into(),
-    ])?
-    .run()
-    .await?;
-
-    Ok(Organization {
-        id: org_id,
-        name: org_name.to_string(),
-        slug,
-        created_at: now,
-        created_by: user_id.to_string(),
-        billing_account_id: Some(billing_account.id),
-    })
-}
-
-/// Get user by ID
-pub async fn get_user_by_id(db: &D1Database, user_id: &str) -> Result<Option<User>> {
-    let stmt = db.prepare(
-        "SELECT id, email, name, avatar_url, oauth_provider, oauth_id, org_id, role, created_at,
-                suspended_at, suspension_reason, suspended_by
-         FROM users
-         WHERE id = ?1",
-    );
-
-    stmt.bind(&[user_id.into()])?.first::<User>(None).await
-}
-
-/// Get organization by ID
-pub async fn get_org_by_id(db: &D1Database, org_id: &str) -> Result<Option<Organization>> {
-    let stmt = db.prepare(
-        "SELECT id, name, slug, created_at, created_by, billing_account_id
-         FROM organizations
-         WHERE id = ?1",
-    );
-
-    stmt.bind(&[org_id.into()])?
-        .first::<Organization>(None)
-        .await
-}
-
-/// Get the org-level forward_query_params default setting (0 or 1).
-/// Returns false if the org doesn't exist or the column is NULL.
-pub async fn get_org_forward_query_params(db: &D1Database, org_id: &str) -> Result<bool> {
-    let stmt = db.prepare(
-        "SELECT COALESCE(forward_query_params, 0) as forward_query_params
-         FROM organizations
-         WHERE id = ?1",
-    );
-
-    let result = stmt
-        .bind(&[org_id.into()])?
-        .first::<serde_json::Value>(None)
-        .await?;
-
-    Ok(result
-        .and_then(|v| v["forward_query_params"].as_f64())
-        .map(|v| v != 0.0)
-        .unwrap_or(false))
-}
-
-/// Get a link by ID
-pub async fn get_link_by_id(db: &D1Database, link_id: &str, org_id: &str) -> Result<Option<Link>> {
-    let stmt = db.prepare(
-        "SELECT id, org_id, short_code, destination_url, title, created_by, created_at, updated_at, expires_at, status, click_count, utm_params, forward_query_params, redirect_type
-         FROM links
-         WHERE id = ?1
-         AND org_id = ?2
-         AND status IN ('active', 'disabled')"
-    );
-
-    stmt.bind(&[link_id.into(), org_id.into()])?
-        .first::<Link>(None)
-        .await
-}
-
-/// Get a link by ID without org_id check (used for admin operations - returns all statuses)
-/// Note: tags are populated separately via get_tags_for_links
-pub async fn get_link_by_id_no_auth_all(db: &D1Database, link_id: &str) -> Result<Option<Link>> {
-    let stmt = db.prepare(
-        "SELECT id, org_id, short_code, destination_url, title, created_by, created_at, updated_at, expires_at, status, click_count, utm_params, forward_query_params, redirect_type
-         FROM links
-         WHERE id = ?1"
-    );
-
-    stmt.bind(&[link_id.into()])?.first::<Link>(None).await
-}
-
-/// Get an active link by short_code without org_id check (used for public reporting)
-/// Only returns links with status = 'active'
-pub async fn get_active_link_by_short_code(
-    db: &D1Database,
-    short_code: &str,
-) -> Result<Option<Link>> {
-    let stmt = db.prepare(
-        "SELECT id, org_id, short_code, destination_url, title, created_by, created_at, updated_at, expires_at, status, click_count, utm_params, forward_query_params, redirect_type
-         FROM links
-         WHERE short_code = ?1
-         AND status = 'active'"
-    );
-
-    stmt.bind(&[short_code.into()])?.first::<Link>(None).await
-}
 
 // UserWithBillingInfo struct moved to repositories/user_repository.rs
 
@@ -572,56 +271,6 @@ pub async fn get_billing_account_for_org(
         .await
 }
 
-/// Get billing account for a user (their primary billing account)
-/// Returns the billing account of the user's primary organization
-pub async fn get_user_billing_account(
-    db: &D1Database,
-    user_id: &str,
-) -> Result<Option<BillingAccount>> {
-    let stmt = db.prepare(
-        "SELECT ba.id, ba.owner_user_id, ba.tier, ba.provider_customer_id, ba.created_at
-         FROM billing_accounts ba
-         INNER JOIN organizations o ON o.billing_account_id = ba.id
-         INNER JOIN users u ON u.org_id = o.id
-         WHERE u.id = ?1",
-    );
-    stmt.bind(&[user_id.into()])?
-        .first::<BillingAccount>(None)
-        .await
-}
-
-/// Create a billing account
-pub async fn create_billing_account(
-    db: &D1Database,
-    owner_user_id: &str,
-    tier: &str,
-) -> Result<BillingAccount> {
-    let id = BillingAccount::generate_id();
-    let now = now_timestamp();
-
-    let stmt = db.prepare(
-        "INSERT INTO billing_accounts (id, owner_user_id, tier, created_at)
-         VALUES (?1, ?2, ?3, ?4)",
-    );
-
-    stmt.bind(&[
-        id.clone().into(),
-        owner_user_id.into(),
-        tier.into(),
-        (now as f64).into(),
-    ])?
-    .run()
-    .await?;
-
-    Ok(BillingAccount {
-        id,
-        owner_user_id: owner_user_id.to_string(),
-        tier: tier.to_string(),
-        provider_customer_id: None,
-        created_at: now,
-    })
-}
-
 /// Get monthly counter for billing account
 pub async fn get_monthly_counter_for_billing_account(
     db: &D1Database,
@@ -694,19 +343,6 @@ pub async fn reset_monthly_counter_for_billing_account(
          WHERE billing_account_id = ?1 AND year_month = ?2",
     );
     stmt.bind(&[billing_account_id.into(), year_month.into()])?
-        .run()
-        .await?;
-    Ok(())
-}
-
-/// Update billing account owner
-pub async fn update_billing_account_owner(
-    db: &D1Database,
-    billing_account_id: &str,
-    new_owner_id: &str,
-) -> Result<()> {
-    let stmt = db.prepare("UPDATE billing_accounts SET owner_user_id = ?1 WHERE id = ?2");
-    stmt.bind(&[new_owner_id.into(), billing_account_id.into()])?
         .run()
         .await?;
     Ok(())
